payment_service/requirements.txt
flask
PyJWT
mysql-connector-python==9.1.0
kafka-python>=2.0.2
